fix compile error on esp32

+++ src/StepperISR.h
@@ -83,8 +83,8 @@ class StepperQueue {
   bool dirHighCountsUp;
   uint8_t dirPin;
 #if defined(ARDUINO_ARCH_ESP32)
-  volatile uint8_t* _dirPinPort;
-  uint8_t _dirPinMask;
+  volatile uint32_t* _dirPinPort;
+  uint32_t _dirPinMask;
   volatile bool _hasISRactive;
   bool isRunning();
   const struct mapping_s* mapping;
